Load fonts once and cache image
Note: Before this we were reloading the HTML page, fonts, and img
for every thumbnail.

With this change we load the image once. But, it was hacked in.
Really it needs to load fonts and images and then cache them but
as it is it only looks at the image and fonts the first time
ThumbnailGenerator.generate is called. That works for this use case
but it's bad code.

But, it's way faster than it was.