Switch from a very old version of marked to markdown-it
I'm not really sure I why I made this change. The old version
worked fine. npm complained their were vulnerabilities in the
old marked but those were irrelevent as this is a static site
builder.

One reason was I'd like to use
[footnotes](https://github.blog/changelog/2021-09-30-footnotes-now-supported-in-markdown-fields/)
and that requires using a newer markdown parser.

Anyway, it seemed like a good idea but took a ton of time.
The biggest issue is since marked shipped 9 years ago
markdown got standardized and IMO made a frustrating choice.
The choice they made is that embedded html ends at the first
blank line. In other words if you have markdown like this

```md

Some paragraph of text.

<div style="background: red">
  <p>
   This is some callout text

   and more text
  </p>
  <p>And some more here</p>
</div>

More normal markdown text.
```

It will see that as 2 chuncks of HTML, not one. In other words it will
generate this HTML

```html
<h1>Title of article</h1>
<p>Some paragraph of text.</p>
<div style="background: red">
  <p>
   This is some callout text
<p>and more text</p>
  </p>
  <p>And some more here</p>
</div>
<p>More normal markdown text.</p>
```

I think I get it. The problem with requiring a markdown parser
to handle the case above is it actually has to be able to parse
HTML to find the closing tag where as what they wanted was to be
able to parse with very simple regular expression rules.

That's fine but for me I just wanted to be able to embed
random HTML in the middle of a document for diagrams or
svg or whatever and not have to remember, ohhhhhhhhh, it's
blank line sensitive.

Further, this builder is already running on files that have
these blank lines.

The old version of marked handled this so how. The new version I'm not
sure but given everyone is moving to [the standard](https://commonmark.org)
most implemenations no longer handle this case.

Anyway, at the moment I have a solution that seems to be working
and unfortunately it does require a full HTML parser. At a glance
it seems to be working on existing content.

I don't know. Maybe I should switch to standard and fix existing
content. Or maybe I should keep existing content old an old version
of the builder.

In anycase, at the moment, this seems to be working.